agents/adk-agent-pocs/mcp_servers/billy/pyproject.toml
[project]
name = "billy-mcp-server"
version = "0.1.0"
description = "Shared FastMCP server exposing Billy accounting stubs for all agents"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "fastmcp>=2.13.0",
    "python-dotenv>=1.2.1",
    "boto3>=1.42.59",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "pytest-mock>=3.14.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
